[k1b-cluster] Feature: Event Interface
