+++ include/arch/cluster/k1b-cluster.h
@@ -45,6 +45,7 @@
 /**@{*/
 
 	#include <arch/cluster/k1b-cluster/cores.h>
+	#include <arch/cluster/k1b-cluster/event.h>
 	#include <arch/cluster/k1b-cluster/memory.h>
 
 	/**
